UnstableIndicator moved to package helper
