+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/UnstableIndicator.java
@@ -21,49 +21,43 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators;
+package org.ta4j.core.indicators.helpers;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotEquals;
-
-import java.util.function.Function;
-
-import org.junit.Before;
-import org.junit.Test;
 import org.ta4j.core.Indicator;
-import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
-import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.indicators.CachedIndicator;
 import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 
-public class UnstableIndicatorTest extends AbstractIndicatorTest<Indicator<Num>, Num> {
+/**
+ * Indicator that returns {@link NaN} in unstable bars.
+ */
+public class UnstableIndicator extends CachedIndicator<Num> {
 
-    private int unstableBars;
-    private UnstableIndicator unstableIndicator;
+    private final int unstableBars;
+    private final Indicator<Num> indicator;
 
-    public UnstableIndicatorTest(Function<Number, Num> numFunction) {
-        super(numFunction);
+    /**
+     * Constructor.
+     * 
+     * @param indicator    the indicator
+     * @param unstableBars the number of first bars of the barSeries to be unstable
+     */
+    public UnstableIndicator(Indicator<Num> indicator, int unstableBars) {
+        super(indicator);
+        this.indicator = indicator;
+        this.unstableBars = unstableBars;
     }
 
-    @Before
-    public void setUp() {
-        unstableBars = 5;
-        unstableIndicator = new UnstableIndicator(new ClosePriceIndicator(new MockBarSeries(numFunction)),
-                unstableBars);
-    }
-
-    @Test
-    public void indicatorReturnsNanBeforeUnstableBars() {
-        for (int i = 0; i < unstableBars; i++) {
-            assertEquals(unstableIndicator.getValue(i), NaN.NaN);
+    @Override
+    protected Num calculate(int index) {
+        if (index < unstableBars) {
+            return NaN.NaN;
         }
+        return indicator.getValue(index);
     }
 
-    @Test
-    public void indicatorNotReturnsNanAfterUnstableBars() {
-        for (int i = unstableBars; i < 10; i++) {
-            assertNotEquals(unstableIndicator.getValue(i), NaN.NaN);
-        }
+    @Override
+    public int getUnstableBars() {
+        return unstableBars;
     }
-
 }
